Fix signed integer overflow in halving tests and Qt parse() MoneyRange gap
Two bugs from the line-by-line audit:

1. validation_tests.cpp + GetBlockSubsidy signature (int overflow)
   With nSubsidyHalvingInterval = 1,042,600,000, multiplying by nHalvings
   as plain int overflows INT32_MAX at nHalvings=3 (3×1,042,600,000 =
   3,127,800,001 > 2,147,483,647), producing undefined behaviour.  Fix by
   widening GetBlockSubsidy's nHeight parameter to int64_t throughout
   (validation.h, validation.cpp) and casting the test arithmetic to
   int64_t in validation_tests.cpp (lines 28 and 34).

2. burritocoinunits.cpp parse() missing MoneyRange check
   The Qt amount parser rejected strings > 19 chars to avoid int64
   overflow, but did not verify the parsed CAmount was within
   [0, MAX_MONEY].  Values between MAX_MONEY+1 and INT64_MAX passed
   silently.  Add an explicit MoneyRange guard before writing val_out.

+++ src/test/validation_tests.cpp
@@ -25,13 +25,13 @@ static void TestBlockSubsidyHalvings(const Consensus::Params& consensusParams)
     CAmount nPreviousSubsidy = nInitialSubsidy * 2;
     for (int nHalvings = 0; nHalvings < maxHalvings; nHalvings++) {
         // Add 1 to avoid height 0 (genesis premine special case).
-        int nHeight = nHalvings * consensusParams.nSubsidyHalvingInterval + 1;
+        int64_t nHeight = (int64_t)nHalvings * consensusParams.nSubsidyHalvingInterval + 1;
         CAmount nSubsidy = GetBlockSubsidy(nHeight, consensusParams);
         BOOST_CHECK(nSubsidy <= nInitialSubsidy);
         BOOST_CHECK_EQUAL(nSubsidy, nPreviousSubsidy / 2);
         nPreviousSubsidy = nSubsidy;
     }
-    BOOST_CHECK_EQUAL(GetBlockSubsidy(maxHalvings * consensusParams.nSubsidyHalvingInterval + 1, consensusParams), 0);
+    BOOST_CHECK_EQUAL(GetBlockSubsidy((int64_t)maxHalvings * consensusParams.nSubsidyHalvingInterval + 1, consensusParams), 0);
 }
 
 static void TestBlockSubsidyHalvings(int nSubsidyHalvingInterval)
